requirements.txt
PyYAML
aiohttp
pillow
discord.py